% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{recaptcha}
\alias{recaptcha}
\title{Server-side module component}
\usage{
recaptcha(input, output, session, secret = Sys.getenv("recaptcha_secret"))
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{secret}{Google secret for reCAPTCHA}
}
\description{
Server-side module component
}
