% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{recaptcha}
\alias{recaptcha}
\title{Server-side module component}
\usage{
recaptcha(id, secret)
}
\arguments{
\item{secret}{Google secret for reCAPTCHA}

\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
Server-side module component
}
