% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{recaptchaUI}
\alias{recaptchaUI}
\title{Client-side module component}
\usage{
recaptchaUI(id, sitekey = Sys.getenv("recaptcha_sitekey"))
}
\arguments{
\item{id}{unique id for module}

\item{sitekey}{Google sitekey code for reCAPTCHA}
}
\description{
Client-side module component
}
